8662: Support separate runtime constraints for individual components.
This improves the cost-efficiency of Arvados pipelines by letting each
component run on an appropriate node size, rather than the node size
required to accommodate the largest component.

This is a flag day change.  When a pipeline is upgraded to this
version of the Queue integration, it must adjust its runtime
components as well.  This was discussed on IRC and deemed worth the
hassle, just because the previously behavior can be so costly.